refactor: move some tests and test mocks into test folder
