+++ src/setupTests.ts
@@ -1,7 +1,7 @@
 import '@testing-library/jest-dom';
 import '@testing-library/jest-dom/vitest';
 import { cleanup } from '@testing-library/react';
-import { server } from './mocks/server';
+import { server } from './tests/mocks/server';
 
 beforeEach(() => {
   server.listen({ onUnhandledRequest: 'error' });
